Move allowedProperties into FormMixin
