+++ viewer/v2client/src/components/inspector/toolbar.vue
@@ -16,11 +16,12 @@ import MarcPreview from '@/components/inspector/marc-preview';
 import FieldAdder from '@/components/inspector/field-adder';
 import TooltipComponent from '@/components/shared/tooltip-component';
 import LensMixin from '@/components/mixins/lens-mixin';
+import FormMixin from '@/components/mixins/form-mixin';
 import { mixin as clickaway } from 'vue-clickaway';
 import { mapGetters } from 'vuex';
 
 export default {
-  mixins: [clickaway, LensMixin],
+  mixins: [clickaway, LensMixin, FormMixin],
   props: {
     fieldAdderOpen: false,
   },
@@ -215,6 +216,17 @@ export default {
       'settings',
       'status',
     ]),
+    formObj() {
+      return this.inspector.data[this.inspector.status.focus];
+    },
+    allowed() {
+      return VocabUtil.getPropertiesFromArray(
+        this.formObj['@type'],
+        this.resources.vocabClasses,
+        this.resources.vocabProperties,
+        this.resources.context
+      );
+    },
     recordType() {
       return VocabUtil.getRecordType(
         this.inspector.data.mainEntity['@type'], 
@@ -261,49 +273,6 @@ export default {
     },
     hasLocalWork() {
       return (typeof this.inspector.data.work !== 'undefined') ? true : false;
-    },
-    allowedProperties() {
-      const settings = this.settings;
-      const formObj = this.inspector.data[this.inspector.status.focus];
-      const allowed = VocabUtil.getPropertiesFromArray(
-        [StringUtil.convertToVocabKey(StringUtil.convertToBaseUri(formObj['@type'], this.resources.context), this.resources.context)],
-        this.resources.vocabClasses,
-        this.resources.vocabProperties,
-        this.resources.context
-      );
-      // Add the "added" property
-      for (const element of allowed) {
-        const oId = StringUtil.getCompactUri(element.item['@id'], this.resources.context);
-        element.added = (formObj.hasOwnProperty(oId) && formObj[oId] !== null);
-      }
-
-      const extendedAllowed = allowed.map(property => {
-        const labelByLang = property.item.labelByLang;
-        if (typeof labelByLang !== 'undefined') {
-          // Try to get the label in the preferred language
-          let label = ((typeof labelByLang[this.settings.language] !== 'undefined') ? labelByLang[this.settings.language] : labelByLang.en);
-          // If several labels are present, use the first one
-          if (_.isArray(label)) {
-            label = label[0];
-          }
-          return {
-            added: property.added,
-            item: property.item,
-            label: label
-          };
-        } else {
-          // If no label, use @id as label
-          return {
-            added: property.added,
-            item: property.item,
-            label: property.item['@id']
-          };
-        }
-      });
-      const sortedAllowed = _.sortBy(extendedAllowed, (prop) => {
-        return prop.label.toLowerCase();
-      });
-      return sortedAllowed;
     },
   },
   components: {
